lookfor_shared_identifier(): Removed wrong cast.
This fun casted a hash in an ident_t to short. This was anyway an
unsigned short but is nowadays a hash16_t which does not need to be
casted at all at that point.
(lex.c)

git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2971 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/lex.c
@@ -1803,7 +1803,7 @@ lookfor_shared_identifier (const char *s, size_t len, int n, int depth, Bool bCr
                         curr->next = inferior->next;
                         curr->type = I_TYPE_UNKNOWN;
                         curr->inferior = inferior;
-                        curr->hash = (short)h;
+                        curr->hash = h;
                         ident_table[h] = curr;
                     }
                 }
